database and seeding

+++ app/database/migrations/setup/2015_01_06_182828_create_taggables_table.php
@@ -32,7 +32,10 @@ class CreateTaggablesTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('taggables');
+		if(Schema::hasTable('taggables')) 
+		{
+			Schema::drop('taggables');
+		}
 	}
 
 }
